Moving functions:machine and customer files
Moved writing functions from customer to
machine. Anything related specifically
to actions taken by the machine should
stay in machine file for better reference
later.
