+++ include/machine.h
@@ -4,5 +4,6 @@
 #include <iostream>
 
 void welcome_menu();
-    
+int write_user_data(std::string first, std::string last, int acct_num, int pin);
+
 #endif
