+++ src/machine.cpp
@@ -2,6 +2,7 @@
 #include "../include/transactions.h"
 
 #include <iostream>
+#include <fstream>
 
 void welcome_menu(){
     int selection;
@@ -36,4 +37,21 @@ void welcome_menu(){
 
     } while (selection <= 5);
 
+}
+
+int write_user_data(std::string first, std::string last, int acct_num, int pin){
+    //write customer data to txt file
+    // account number, first, last, balance, pin
+    std::ofstream outfile;
+
+    outfile.open("../data/accounts.txt", std::ios::app); // open file
+    if (outfile.is_open()) //confirm file opened
+    {
+        outfile << acct_num;
+        outfile.close();
+        
+    }
+    else
+        std::cerr << "Error writing to database";
+    
 }
